refactor: convert `InputMedia` to an interface and add `Uploadable` implementation for all input media types
