+++ src/Telegram/Types/Input/InputMediaLivePhoto.php
@@ -8,7 +8,9 @@ use SergiX44\Hydrator\Annotation\SkipConstructor;
 use SergiX44\Hydrator\Resolver\EnumOrScalar;
 use SergiX44\Nutgram\Telegram\Properties\InputMediaType;
 use SergiX44\Nutgram\Telegram\Properties\ParseMode;
+use SergiX44\Nutgram\Telegram\Types\BaseType;
 use SergiX44\Nutgram\Telegram\Types\Internal\InputFile;
+use SergiX44\Nutgram\Telegram\Types\Internal\Uploadable;
 use SergiX44\Nutgram\Telegram\Types\Message\MessageEntity;
 use function SergiX44\Nutgram\Support\array_filter_null;
 
@@ -17,11 +19,29 @@ use function SergiX44\Nutgram\Support\array_filter_null;
  * @see https://core.telegram.org/bots/api#inputmedialivephoto
  */
 #[SkipConstructor]
-class InputMediaLivePhoto extends InputMedia implements InputPollMedia, InputPollOptionMedia, JsonSerializable
+class InputMediaLivePhoto extends BaseType implements InputMedia, InputPollMedia, InputPollOptionMedia, Uploadable, JsonSerializable
 {
     /** Type of the result, must be live_photo */
     #[EnumOrScalar]
     public InputMediaType|string $type = InputMediaType::LIVE_PHOTO;
+
+    /**
+     * File to send.
+     * Pass a file_id to send a file that exists on the Telegram servers (recommended),
+     * pass an HTTP URL for Telegram to get a file from the Internet,
+     * or pass “attach://<file_attach_name>” to upload a new one using multipart/form-data under <file_attach_name> name.
+     * {@see https://core.telegram.org/bots/api#sending-files More information on Sending Files »}
+     */
+    public InputFile|string $media;
+
+    /**
+     * The static photo to send.
+     * Pass a file_id to send a file that exists on the Telegram servers (recommended),
+     * pass an HTTP URL for Telegram to get a file from the Internet,
+     * or pass “attach://<file_attach_name>” to upload a new one using multipart/form-data under <file_attach_name> name.
+     * {@see https://core.telegram.org/bots/api#sending-files More information on Sending Files »}
+     */
+    public InputFile|string $photo;
 
     /**
      * Optional.
@@ -58,6 +78,7 @@ class InputMediaLivePhoto extends InputMedia implements InputPollMedia, InputPol
 
     public function __construct(
         InputFile|string $media,
+        InputFile|string $photo,
         ?string $caption = null,
         ParseMode|string|null $parse_mode = null,
         ?array $caption_entities = null,
@@ -66,6 +87,7 @@ class InputMediaLivePhoto extends InputMedia implements InputPollMedia, InputPol
     ) {
         parent::__construct();
         $this->media = $media;
+        $this->photo = $photo;
         $this->caption = $caption;
         $this->parse_mode = $parse_mode;
         $this->caption_entities = $caption_entities;
@@ -75,6 +97,7 @@ class InputMediaLivePhoto extends InputMedia implements InputPollMedia, InputPol
 
     public static function make(
         InputFile|string $media,
+        InputFile|string $photo,
         ?string $caption = null,
         ParseMode|string|null $parse_mode = null,
         ?array $caption_entities = null,
@@ -83,6 +106,7 @@ class InputMediaLivePhoto extends InputMedia implements InputPollMedia, InputPol
     ): self {
         return new self(
             media: $media,
+            photo: $photo,
             caption: $caption,
             parse_mode: $parse_mode,
             caption_entities: $caption_entities,
@@ -97,11 +121,27 @@ class InputMediaLivePhoto extends InputMedia implements InputPollMedia, InputPol
         return array_filter_null([
             'type' => $this->type,
             'media' => $this->media,
+            'photo' => $this->photo,
             'caption' => $this->caption,
             'parse_mode' => $this->parse_mode,
             'caption_entities' => $this->caption_entities,
             'show_caption_above_media' => $this->show_caption_above_media,
             'has_spoiler' => $this->has_spoiler,
         ]);
+    }
+
+    public function isLocal(): bool
+    {
+        return $this->media instanceof InputFile;
+    }
+
+    public function getResource()
+    {
+        return $this->media->getResource();
+    }
+
+    public function getFilename(): string
+    {
+        return $this->media->getFilename();
     }
 }
